+++ tonl.ps1
@@ -1,8 +1,0 @@
-#!/usr/bin/env pwsh
-param(
-    [Parameter(ValueFromRemainingArguments)]
-    [string[]]$Arguments
-)
-
-$cliPath = Join-Path $PSScriptRoot "dist\cli.js"
-& node $cliPath $Arguments